Add some tools from osdk repo
These tools are not meant to be used as-is They need a full rework.
They are more reference material than anything else right now. I
also had to do some awful stuff to them to get them to pass the gate
in a hurry, code looks bad because of it.

I am cleaning them up as I go.

We will remove these in the future once they are no longer useful to
reference any longer.

Change-Id: I8f13ef282a2a05e72c53a0c307832a46db18d2b4