[WIP] Add manifest parser
This reads the manifest. Its more based on osdk.py that I initially
thought it would be. But hopefully I have cleaned it up a good bit.

Update backup.py to take advantage of this.

To generate a manifest with 2^24 objects (64TB for 4MB objects) run:
python backup.py --fakeblockdev 64000 --manifest /pathtosave/manifest

Change-Id: I09a8fb64c0d4b8e4b4b436cca44b57bec87116cd