Create manifest and implement sqlite driver
Use class inheritance for driver in case we need something other than
sqlite as time goes on.

SQLite is ~2x bigger than the custom format used with osdk. When
compressed the SQLite manifest is only about 50% bigger than osdk.
The benefit we get from using common contructs (sqlite) and tracking
and changing the manifest schema is enough for me to think this is
right way to proceed.

To generate a manifest with 2^24 objects (64TB disk) run:
tools/generate_manifest.py --backupsize 64000 --manifest /pathtosave/manifest

This will generate a 809MB file

Co-Authored-By: Sam Yaple <sam@yaple.net>
Change-Id: Ic431bfa52b6fcaeb1c6a64cf270cbb36c496335e